Make the jl::MaybeQuoted I/O manipulator safer
By having it verify that the entire string is properly quoted before
passing it along unmodified. Because having properly quoted strings are
important, and the given string could be malicious.

This is more compute intensive (properly negligible for small-ish
strings) than assuming that a string that looks quoted is quoted, but
security is kind of important nowadays, so it is way better to do it
properly than to have an exploitable version that is faster.

Also add useful jl::find_unescaped and jl::needs_quotes helper functions
to deal with quoted strings and escape sequences.