Replace jl::sliced with jl::chunked
Change the name to align with c++23 std::ranges::views::chunk, and
reimplement as a random access range over the span instead of a
materialized vector of spans.

Also make it more compatible with views::chunk by allowing the last span
to be smaller than the rest instead of asserting that they divide the
chunk size.